[1192] Author: phpnut Date: 4:50:40 AM, Monday, October 24, 2005 Message: Adding comment to doc block with possible settings for the driver configuration
[1191]
Author: phpnut
Date: 4:37:46 AM, Monday, October 24, 2005
Message:
Adding configuration setting in database.php to set the type of connection you want to use.
Read the comments in the doc block to explain how to set the connection type.
doc block also has information on setting the port to connect to.
Corrected Html::tableHeaders();
Closing Ticket #98 and Ticket #26

[1190]
Author: phpnut
Date: 3:25:00 AM, Monday, October 24, 2005
Message:
Fix for Ticket #36  


git-svn-id: https://svn.cakephp.org/repo/trunk/cake/cake@1193 3807eeeb-6ff5-0310-8944-8be069107fe0