Merging fixes and enhancements into trunk
Revision: [1848]
Fixing references.

Revision: [1847]
Adding fix for errors when using scaffold.
On some version of PHP 4 some people where getting Call to member function on non-object...
in Controller::generateFieldNames() at line 494.

Fixed call to get_class_methods that was using the name of the class and not the instance of the class,
this would cause a Wrong datatype for second argument...


git-svn-id: https://svn.cakephp.org/repo/trunk/cake/cake@1849 3807eeeb-6ff5-0310-8944-8be069107fe0