Merging fixes and enhancements into trunk
Revision: [1907]
Fixed doc comment for loadPluginModels()

Revision: [1906]
Added function to load plugin models.
This will not attempt to load an AppModel used by a plugin.
More work is needed on work around for this.
A conflict with class names would be an issue.

Revision: [1905]
Adding component callbacks and RequestHandler component

git-svn-id: https://svn.cakephp.org/repo/trunk/cake/cake@1908 3807eeeb-6ff5-0310-8944-8be069107fe0