Merging fixes and enhancements into trunk.
Revision: [2257]
Refactored cache code in view class.
Removed adding the timestamp to the file contents, it is now
added to the file name

Revision: [2256]
Refactored cache checking code.
Time stamp is not added to file name so we check this 
instead of reading the file and finding timestamp

Revision: [2255]
Adding clearCache function to delete contents of cache

git-svn-id: https://svn.cakephp.org/repo/trunk/cake/cake@2258 3807eeeb-6ff5-0310-8944-8be069107fe0